+++ FirebaseAI/CHANGELOG.md
@@ -1,4 +1,5 @@
 # Unreleased
+- [added] Added support for Grounding with Google Search. (#15014)
 - [removed] Removed `CountTokensResponse.totalBillableCharacters` which was
   deprecated in 11.15.0. Use `totalTokens` instead. (#15056)